refactor: cleaned up unused imports
Also moved some internal functions into more sensible places now that
the app shell has been forked into subsequent/initial loads.
