+++ packages/perseus/src/utils/mod.rs
@@ -9,12 +9,16 @@ mod decode_time_str;
 mod log;
 mod path_prefix;
 mod test;
+#[cfg(target_arch = "wasm32")]
+mod fetch;
 
 pub(crate) use async_fn_trait::AsyncFnReturn;
 #[cfg(not(target_arch = "wasm32"))]
 pub use cache_res::{cache_fallible_res, cache_res};
 #[cfg(target_arch = "wasm32")]
 pub use checkpoint::checkpoint;
+#[cfg(target_arch = "wasm32")]
+pub(crate) use fetch::fetch;
 pub(crate) use context::provide_context_signal_replace;
 #[cfg(not(target_arch = "wasm32"))]
 pub use decode_time_str::{ComputedDuration, Duration, InvalidDuration};
